backend2/requirements.txt
fastapi==0.108.0
SQLAlchemy==2.0.40
pandas==2.1.3
uvicorn==0.30.6
PyJWT==2.10.1
